requirements.txt
peewee
openai
python-dotenv
pymysql
cryptography
json-repair